backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
networkx
numpy
scikit-learn
openai
httpx

# Testing dependencies
pytest
pytest-asyncio
